symlinker script actually works now wowgit add .

+++ symlinker.sh
@@ -4,27 +4,34 @@
 
 files=(".zshrc" ".vimrc" ".gitconfig" ".gitignore_global" "Brewfile")
 dirs=(".vim") 
-if [ ! -d $dir ]; then    
-    mkdir ~/.old_dotfiles
+
+cd; echo "moving to home directory"
+
+if [ ! -d .old_dotfiles ]; then    
+    mkdir .old_dotfiles
+else
+    rm -r ".old_dotfiles"
+    mkdir .old_dotfiles
 fi
 
-echo "moving to .dotfiles dir"
-cd ~/.dotfiles
-echo "done"
+echo "created old_dotfiles dir"
 
 # making directories that need symlinked later
 for dir in ${dirs[@]}; do
-    if [ ! -d $dir ]; then    
-        mkdir ~/.dotfiles/$dir 
-    fi    
-    mv ~/$dir ~/.old_dotfiles
+    if [ -d $dir ]; then    
+        echo "attempting to move existing dir $dir to ~/.old_dotfiles"
+        mv ~/$dir ~/.old_dotfiles
+    fi
+    echo "symlinking $dir"
     ln -s ~/.dotfiles/$dir ~/$dir
 done
 
 # symlinking all files 
 for file in ${files[@]}; do
-    echo "attempting to move existing file $file to ~/.old_dotfiles"
-    mv ~/$file ~/.old_dotfiles
+    if [ -f $file ]; then
+        echo "attempting to move existing file $file to ~/.old_dotfiles"
+        mv ~/$file ~/.old_dotfiles
+    fi
     echo "symlinking $file"
     ln -s ~/.dotfiles/$file ~/$file
 done
